requirements.txt
tzlocal==1.5.1
pandas==0.25.1
numpy==1.17.2
matplotlib==3.1.1
pillow==6.2.0
streamlit==0.57.2
textract==1.6.3
